fix: harden detachable panel behavior
